+++ lib/nanoc/base/services/item_rep_router.rb
@@ -3,6 +3,12 @@ module Nanoc::Int
   #
   # @api private
   class ItemRepRouter
+    class IdenticalRoutesError < ::Nanoc::Error
+      def initialize(output_path, rep_a, rep_b)
+        super("The item representations #{rep_a.inspect} and #{rep_b.inspect} are both routed to #{output_path}.")
+      end
+    end
+
     def initialize(reps, rules_collection, site)
       @reps = reps
       @rules_collection = rules_collection
@@ -10,19 +16,27 @@ module Nanoc::Int
     end
 
     def run
+      paths_to_reps = {}
       @reps.each do |rep|
         rules = @rules_collection.routing_rules_for(rep)
         raise Nanoc::Int::Errors::NoMatchingRoutingRuleFound.new(rep) if rules[:last].nil?
 
         rules.each_pair do |snapshot, rule|
-          route_rep(rep, snapshot, rule)
+          route_rep(rep, snapshot, rule, paths_to_reps)
         end
       end
     end
 
-    def route_rep(rep, snapshot, rule)
+    def route_rep(rep, snapshot, rule, paths_to_reps)
       basic_path = basic_path_for(rep, rule)
       return if basic_path.nil?
+
+      # Check for duplicate paths
+      if paths_to_reps.key?(basic_path)
+        raise IdenticalRoutesError.new(basic_path, paths_to_reps[basic_path], rep)
+      else
+        paths_to_reps[basic_path] = rep
+      end
 
       rep.raw_paths[snapshot] = @site.config[:output_dir] + basic_path
       rep.paths[snapshot] = strip_index_filename(basic_path)