+++ lib/nanoc/cli/commands/prune.rb
@@ -1,12 +1,12 @@
 usage 'prune'
-summary 'remove files not managed by nanoc from the output directory'
+summary 'remove files not managed by Nanoc from the output directory'
 description <<-EOS
 Find all files in the output directory that do not correspond to an item
-managed by nanoc and remove them. Since this is a hazardous operation, an
+managed by Nanoc and remove them. Since this is a hazardous operation, an
 additional `--yes` flag is needed as confirmation.
 
 Also see the `auto_prune` configuration option in `nanoc.yaml` (`config.yaml`
-for older nanoc sites), which will automatically prune after compilation.
+for older Nanoc sites), which will automatically prune after compilation.
 EOS
 
 flag :y, :yes,       'confirm deletion'
@@ -25,7 +25,7 @@ module Nanoc::CLI::Commands
       else
         $stderr.puts 'WARNING: Since the prune command is a destructive command, it requires an additional --yes flag in order to work.'
         $stderr.puts
-        $stderr.puts 'Please ensure that the output directory does not contain any files (such as images or stylesheets) that are necessary but are not managed by nanoc. If you want to get a list of all files that would be removed, pass --dry-run.'
+        $stderr.puts 'Please ensure that the output directory does not contain any files (such as images or stylesheets) that are necessary but are not managed by Nanoc. If you want to get a list of all files that would be removed, pass --dry-run.'
         exit 1
       end
     end