+++ NEWS.md
@@ -1,6 +1,22 @@
-# nanoc news
+# Nanoc news
 
-## ??? (???)
+## 4.0.0rc2 (???)
+
+Fixes:
+
+* Fixed broken `shell` command (#672) [Jim Mendenhall]
+* Fixed absolute path check on Windows (#656)
+
+Enhancements:
+
+* Made Nanoc error when multiple items have the same output path (#665, #669)
+* Improved error message for non-hash frontmatter (#670, #673)
+
+Changes:
+
+* nanoc is now called Nanoc
+
+## 4.0.0rc1 (2015-06-21)
 
 Fixes: